initial error handling for pre-existing chunks
When the database interrogates the chunkservers for their chunk
contents, often in testing, chunks exist on the chunkserver that are
not associated with a file in the oplog, making them orphaned chunks.
This currently allows initialization to continue without crashing in
this event. Ideally in the future, orphaned chunks would be handled
better than via a logging.warning message.