fix for heisenbug that intermittently prevented append
Sometimes, directly after creating a new file, appending raised a
KeyError. This was because the chunkHandle was converted to a string
down the line, so in some places it leaked as an int. Converting it to
an int right when getChunkHandle() is called. Now, everything
downstream of it will be a string, resolving the intermittent Key/Value
Errors